Boss progress
Boss framework written. Mini-boss will act as a larger version of a
regular enemy with the same behavior, just different speed/hp/damage.

Main boss will have 3 phases, 1 where they are shielded and we have to
break an object on screen to remove the shield, 1 where they are moving
around on screen, and one where they sit still and shoot some
projectiles that track the player.