Save/load fixes
So basically, javaFX Image class objects can't be serialized. To get
around this I made an GameImages class that creates the images for each
game object and stores it as a field, then when we need to use one we
can just call it directly from that class.

Past that, the only issue with the load/save function was remembering to
have save writeObject, closing both of the streams when they were done,
and changing how startGame works a little to take in a GameModel and
create a controller based on that. Also, remembering to start the
animation back up after the startgame code since it was still paused
from opening the menu.